arm: the head servo gets its own file
Pure move of the head section out of arm_services.cpp into arm_head.cpp:
different hardware, its own topic and services, no business in a file
about the arm's services.
